Add files via upload
A First draft of importing fuel price data from the source.  A Try except loop was used to get past the occasion where the link to the data had skipped a week. This first draft is extremely slow as the try except is not optimal. the next step is to improve efficiency of the code by trying to optimise the try except loop. I have yet to add comments to this code as this is mainly a test to make sure i have set up Git correctly.